+++ CHANGELOG.md
@@ -1,1 +1,4 @@
-* Fixed default player.json.
+* Fixed afterimages showing in first person when used with Iris.
+* Fixed afterimages rendering over clouds.
+* Changed default config value for step size to be more performant.
+* Fixed incompatibility with EMF.